Cleaned up the logic in 'runastrodriz' to only try to delete the OrIg_files directory if it is empty.  Not strictly needed for pipeline use, since parameters are set in the script to insure that directory never gets made, but it makes the script safer for testing.
git-svn-id: http://svn.stsci.edu/svn/ssb/stsci_python/stsci_python/trunk/acstools@16511 fe389314-cf27-0410-b35b-8c050e845b92

+++ lib/acstools/runastrodriz.py
@@ -283,7 +283,12 @@ def process(inFile,force=False):
 
     # Clean up any generated OrIg_files directory
     if os.path.exists("OrIg_files"):
-        os.rmdir("OrIg_files")
+        # check to see whether this directory is empty
+        flist = glob.glob('OrIg_files/*.fits')
+        if len(flist) == 0:
+            os.rmdir("OrIg_files")
+        else:
+            print 'OrIg_files directory NOT removed as it still contained images...'
 
     # Provide feedback to user
     print _final_msg
